fix(spinbot): Forbid concurrency
Spinnakerbot sometimes runs into github rate limits; given that
it runs every 5 minutes, we we should just let it fail and retry
in 5 minutes. Instead the default config is leading to multiple
retries of the same job and concurrent executions.

I actually already applied these changes directly to the cluster before
I realized we had this manifest in source control, so it's been running
with these changes for a while. This just updates the source to match
what's running.