fix(spinbot): Reduce API usage of label filtering
We're currently calling the API to get an issue's labels every time
we want to check if it has a particular label. The labels are already
on the object we have, so just use issue.labels instead of get_labels.

Also minorly simplify the logic, which looked to be an attempt to
be efficient by short-circuiting the check as soon as a label was
found but ended up writing code that was complex and masked a real
performance issue. (I'd rather iterate over a list that will probably
have a length of 5 at most than make a network call.)