chore(core): Delete monitoring code
The monitoring database for spinnakerbot has not been functional for
several months. Given that we now have spinnaker.devstats.cd.foundation,
the benefit of maintaining our own metrics is not clear (particularly
given that they have not been maintained). These metrics are also
largely responsible for spinnakerbot hitting rate limits.

If at some point in the future we want to re-implement metrics, we
can do so and pull back this code, but given that it's just a really
expensive no-op now and it's not clear that fixing it is a priority,
just deleting the code.