perf(core): Reduce size of docker container and use cached build
We can greatly reduce the size of the spinnakerbot image by using
an alpine base image. This reduces the image size from ~340MB to
~40MB.

Also, move the 'pip install' command before we copy over the whole
app. This greatly speeds up local docker builds as we don't need
to re-install all requirements if they haven't changed since the
last build. Also change the GCB build to cache from the prior build
on the same branch so that it gets the same advantage. (This requires
pulling it out into a cloudbuild.yaml file.)

Remove the chown command; one of the advantages of not running as
root is that the app doesn't have permission to write to its source
code; chown'ing all the source to be owned by the app defeats this.

Finally, change CMD to ENTRYPOINT as we don't need to go through
a shell to execute this python script.