perf(core): Remove iteration over repos and PRs
After removing the metrics code, the only remaining handler looks
at issues only, so remove the iteration over repos and PRs in the
policy code. (Note that we *do* still look at PRs in the event code,
such as to respond to cherry picks and to look for new PRs writing
groovy code---but we don't do anything with the result of listing
all PRs in the repo.)

This removes a bunch of calls to the API where we are just throwing
away the result.