perf(core): Only look for issues in repos that allow issues
Most of our repos don't allow issues, so we're wasting our time
polling every few minutes for issues in these repos. (And it's not
always a cheap call as the issues endpoint also returns PRs which
we filter on the client side.)

Allow the event handler and policy handler to separately configure
which repos to look at; the event handler will look at all repos,
but the policy handler (which now only cares about issues) will
only look at repos with issues enabled.