perf(core): A few more minor perf fixes to spinnakerbot (#10)
* fix(core): Fix some lint warnings

* logging.warn is deprecated; use logging.warning
* The google package was depended on but was not in requirements,
though was being transitively included
* Add venv to .gitignore, as IntelliJ uses this to create a virtual
environment when you set up python support

* perf(core): Remove iteration over repos and PRs

After removing the metrics code, the only remaining handler looks
at issues only, so remove the iteration over repos and PRs in the
policy code. (Note that we *do* still look at PRs in the event code,
such as to respond to cherry picks and to look for new PRs writing
groovy code---but we don't do anything with the result of listing
all PRs in the repo.)

This removes a bunch of calls to the API where we are just throwing
away the result.

* perf(core): Only look for issues in repos that allow issues

Most of our repos don't allow issues, so we're wasting our time
polling every few minutes for issues in these repos. (And it's not
always a cheap call as the issues endpoint also returns PRs which
we filter on the client side.)

Allow the event handler and policy handler to separately configure
which repos to look at; the event handler will look at all repos,
but the policy handler (which now only cares about issues) will
only look at repos with issues enabled.