refactor(bot): remove obsolete spinnakerbot cherry-pick command (#15)
We are now using the mergify backport command to handle automated cherry-picks against release branches. For a couple of releases, we had spinnakerbot print a warning and handle the mergify backport command for users, but that has resulted in users who do not have sufficient permission to run the mertify command themselves having spinnakerbot run it for them.